+++ packages/react-data-structures/src/linkedlist/LinkedList.tsx
@@ -1,0 +1,13 @@
+import React, { useState } from "react";
+
+interface LinkedListProp {
+    children?: React.ReactNode
+}
+
+export const LinkedList = ({ children }: LinkedListProp) => {
+    return (
+        <React.Fragment>
+            {children}
+        </React.Fragment>
+    )
+}
