--- packages/react-data-structures/src/linkedlist/node.css
@@ -1,17 +1,0 @@
-.node-ui-triangle, .node-ui-rectangle  {
-    background: red;
-    height: 30px;
-    width: 60px;
-}
-
-.node-ui-triangle::before {
-    width: 0; 
-    height: 0; 
-    border-top: 10px solid transparent;
-    border-bottom: 10px solid transparent; 
-    border-right:10px solid red; 
-}
-
-.node-ui-rectangle::before {
-
-}
